+++ codex/settings/settings.py
@@ -118,7 +118,8 @@ WSGI_APPLICATION = "codex.wsgi.application"
 # https://docs.djangoproject.com/en/3.0/ref/settings/#databases
 
 DB_PATH = CONFIG_PATH / "db.sqlite3"
-BACKUP_DB_PATH = DB_PATH.with_suffix(DB_PATH.suffix + ".bak")
+BACKUP_DB_DIR = CONFIG_PATH / "backups"
+BACKUP_DB_PATH = (BACKUP_DB_DIR / DB_PATH.stem).with_suffix(DB_PATH.suffix + ".bak")
 
 DATABASES = {
     "default": {